Виправлення в тальнівській програмі сатурації. Поправлено з’єднання сигналів в панелі управління продувкою фільтра-відстійника.
git-svn-id: file:///tmp/qt4/talne@340 76bea514-bf2f-4ad3-ac3e-82dc02406d2c

+++ satur/scada/dlgcessbleding.cpp
@@ -20,8 +20,8 @@ dlgCessBleding::dlgCessBleding(IoDev &source, QWidget *parent) :
 
     connect(ui->Am_17,SIGNAL(currentIndexChanged(int)),this,SLOT(slotSet(int)));
     connect(ui->X_zd_17,SIGNAL(valueChanged(double)),this,SLOT(slotSet(double)));
-    connect(ui->Ton_X_17,SIGNAL(valueChanged(double)),this,SLOT(slotSet(int)));
-    connect(ui->Toff_X_17,SIGNAL(valueChanged(double)),this,SLOT(slotSet(int)));
+    connect(ui->Ton_X_17,SIGNAL(valueChanged(int)),this,SLOT(slotSet(int)));
+    connect(ui->Toff_X_17,SIGNAL(valueChanged(int)),this,SLOT(slotSet(int)));
 
     connect(ui->sX_17,SIGNAL(valueChanged(int)),this,SLOT(slotSet(int)));
     connect(ui->dX_17,SIGNAL(valueChanged(double)),this,SLOT(slotSet(double)));
